Move infobox headers from left to above
On my display, the code is long enough that it wraps, which is ugly.
We could just make the code shorter, but there's a better solution.
The low-resolution responsive view puts the infobox headers *above*,
rather than to the left, which not only wastes less space but also
just plain looks better. So let's just do that on the high-res view.