Add community subteam
Add community subteam, based on the list from [the announcement][1].
While no team lead was specified in the announcement, since Steve
Klabnik was the one to organize and announce it, mark him as the team
lead.

I could not find a GitHub account for Rachel Craig, so I refactored the
avatar and URL code to be able to link to an account on different sites;
right now, Twitter and GitHub are defined.  The easiest way I could find
to embed an avatar from Twitter was via avatars.io, though sadly it come
out at lower resolution than the GitHub avatars that way.

[1]: https://internals.rust-lang.org/t/announcing-the-community-subteam/2248