Optimize the SVG version of the Rust logo
Removed unneeded whitespace, comments and xml stuff.
Changed parts that should be circles to use arcs instead of bezier
curves.
Combined the two bezier curves of the teeth to one through doing de
Casteljau’s algorithm in reverse.
Changed the parts between the teeth to lines as the curves are too
small to be seen by the eyes.
Removed some tiny curves at the lower right end of the R.