Remove claim about "thread safety"
As https://github.com/rust-lang/rust/issues/26215 and others have pointed out, "thread safety" is a bad way to phrase this. When talking to people about Rust, the thing they've found most striking is compile-time errors, which is really unique to us, so it seems like a good proxy here.