Retaining the fragment identifier when redirecting
Currently, if a user requests /team.html#foo and is redirected to
the sv-SE language, the new location will be /sv-SE/team.html.
This commit appends the fragment identifier to the redirection
target. Thus, if a user is requesting /team.html#foo and is
redirected to the sv-SE language, the new location will be
/sv-SE/team.html#foo.