create the "docs peers" subteam on the teams page
For a while now, the docs team has let Havvy manage, approve, and merge PRs to the reference as a sort of "peer" to the docs team, specifically managing the reference. To call out their contributions (and commit access), we'd like to float the idea of "docs peers", who "own" a given piece of docs, with Havvy as its inaugural member.