Fixing link to MS VS 2017 tools for Windows
The french visitors of the *Microsoft Visual Studio website* are redirected on its french version, but when they are redirected, they are loosing the ID of the subsection of the tool they need to install for Microsoft Windows use (the `#build-tools-for-visual-studio-2017` part on the end of the URL). Furthermore, the name of the tool in french was wrong (or has changed since translated).

I've discovered the correct name of the software when I changed the name of the *Microsoft Visual Studio website* to English and click another time on the link.